Added infinite test. Some simulations produce correct data, but the test on the overlap with the TEBD evolved state fails.
